+++ computer_name.txt
@@ -1,1 +1,1 @@
-MEET-EliteBook-840-G2
+meet-HP-EliteBook-840-G3
